+++ server/src/http/controllers/shared.types.ts
